vote-interface: move V3 tests into vote_state_v3.rs (#590)
Move 8 tests that exclusively exercise VoteStateV3 APIs out of the
monolithic mod.rs test module and into vote_state_v3.rs where they
belong: size_of, minimum_balance, serialize, deserialize_into,
deserialize_into_error, deserialize_into_uninit, nopanic, and
ill_sized.
